GH-37: Applied some changes based on the PR review feedbacks.
