+++ .vscode/settings.json
@@ -1,3 +1,8 @@
+// If one would like to add/remove/modify user preferences without modifying the content of the
+// workspace settings file, then one would need to modify the `settings.json` under here:
+// - Windows: %APPDATA%\Code\User\settings.json
+// - Linux: $HOME/.config/Code/User/settings.json
+// - Mac: $HOME/Library/Application Support/Code/User/settings.json
 {
     "tslint.enable": true,
     "tslint.configFile": "./config/tslint/tslint.json"
