+++ src/application/common/context.ts
@@ -1,8 +1,8 @@
-import { PredicateExt } from './predicates';
-
-export interface Context<T> extends PredicateExt<T> {
+export interface Context<T> {
     /**
      * The unique ID of the current context.
      */
     readonly id: string;
+
+    isEnabled(arg?: T): boolean;
 }
