+++ src/monaco/browser/monaco-module.ts
@@ -1,8 +1,10 @@
 import { ContainerModule, injectable, decorate} from "inversify";
 import { Languages, Workspace } from "../../languages/common";
-import { MonacoWorkspace } from './monaco-workspace';
-import { MonacoLanguages } from './monaco-languages';
+import { MonacoWorkspace } from "./monaco-workspace";
+import { MonacoLanguages } from "./monaco-languages";
 import { MonacoToProtocolConverter, ProtocolToMonacoConverter } from "monaco-languageclient";
+import { MonacoKeybindingContribution} from "./monaco-keybinding";
+import { KeybindingContribution } from "../../application/common/keybinding";
 
 decorate(injectable(), MonacoToProtocolConverter);
 decorate(injectable(), ProtocolToMonacoConverter);
@@ -15,6 +17,7 @@ export {
 }
 
 export const monacoModule = new ContainerModule(bind => {
+    bind<KeybindingContribution>(KeybindingContribution).to(MonacoKeybindingContribution);
     bind(MonacoToProtocolConverter).toSelf().inSingletonScope();
     bind(ProtocolToMonacoConverter).toSelf().inSingletonScope();
     bind(Languages).to(MonacoLanguages).inSingletonScope();
